Verify backup publication before retention

+++ resources/lib/backup.py
@@ -18,6 +18,8 @@ from resources.lib.archive import (
     build_manifest,
     load_manifest,
     sha256_reader,
+    verify_manifest_files,
+    verify_zip_archive,
 )
 from resources.lib.planning import (
     FilePlanner,
@@ -68,7 +70,12 @@ class XbmcBackup:
         self.transferSize = 0
         self.transferLeft = 0
         self.backup_plan = None
+        self.backup_manifest = None
+        self.backup_manifest_file_hash = None
         self._automatic_exclusion_rules = None
+        self._active_artifact = None
+        self._active_artifact_compressed = False
+        self._vfs_closed = True
 
         self.configureRemote()
         utils.log(utils.getString(30046))
@@ -130,33 +137,64 @@ class XbmcBackup:
         self.skip_advanced = True
 
     def backup(self, progressOverride=False):
+        try:
+            return self._runBackup(progressOverride)
+        except Exception as error:
+            utils.log('Backup failed: %s' % error, xbmc.LOGWARNING)
+            self._discardActiveArtifact()
+            utils.showNotification(utils.getString(30092))
+            return False
+        finally:
+            if not self._vfs_closed:
+                self._closeVFS()
+
+    def _runBackup(self, progressOverride=False):
         shouldContinue = self._setupVFS(self.Backup, progressOverride)
 
         if(shouldContinue):
             utils.log(utils.getString(30023) + " - " + utils.getString(30016))
-            # check if remote path exists
-            if(self.remote_vfs.exists(self.remote_vfs.root_path)):
-                # may be data in here already
-                utils.log(utils.getString(30050))
-            else:
-                # make the remote directory
-                self.remote_vfs.mkdir(self.remote_vfs.root_path)
+            # Folder backups must never reuse a prior timestamp root. ZIP
+            # archives create their logical root inside the local ZIP instead.
+            if not isinstance(self.remote_vfs, ZipFileSystem):
+                if self.remote_vfs.exists(self.remote_vfs.root_path):
+                    utils.log('Backup target already exists: ' +
+                              self.remote_vfs.root_path, xbmc.LOGWARNING)
+                    utils.showNotification(utils.getString(30092))
+                    self._closeVFS()
+                    return False
+                if not self.remote_vfs.mkdir(self.remote_vfs.root_path):
+                    utils.log('Unable to create backup target: ' +
+                              self.remote_vfs.root_path, xbmc.LOGWARNING)
+                    utils.showNotification(utils.getString(30092))
+                    self._closeVFS()
+                    return False
+                self._active_artifact = self.remote_vfs.root_path
+                self._active_artifact_compressed = False
 
             utils.log(utils.getString(30051))
             utils.log('File Selection Type: ' + str(utils.getSetting('backup_selection_type')))
             allFiles = self._collectBackupFiles()
 
-            # create a validation file for backup rotation
-            writeCheck = self._createValidationFile(allFiles)
+            try:
+                writeCheck = self._createValidationFile(allFiles)
+            except Exception as error:
+                utils.log('Unable to create Backup Pro manifest: %s' % error,
+                          xbmc.LOGWARNING)
+                writeCheck = False
 
             if(not writeCheck):
-                # we may not be able to write to this destination for some reason
-                shouldContinue = xbmcgui.Dialog().yesno(utils.getString(30089), "%s\n%s" % (utils.getString(30090), utils.getString(30044)), autoclose=25000)
-
-                if(not shouldContinue):
-                    return
+                if isinstance(self.remote_vfs, ZipFileSystem):
+                    utils.showNotification(utils.getString(30092))
+                else:
+                    self._finalizeBackup(
+                        False, self.remote_vfs.root_path, compressed=False)
+                self._closeVFS()
+                return False
 
             orig_base_path = self.remote_vfs.root_path
+            backup_success = True
+            compressing = utils.getSettingBool("compress_backups")
+            remote_artifact = None if compressing else orig_base_path
 
             # backup all the files
             self.transferLeft = self.transferSize
@@ -166,43 +204,83 @@ class XbmcBackup:
                 filesCopied = self._copyFiles(fileGroup['files'], self.xbmc_vfs, self.remote_vfs)
 
                 if(not filesCopied):
-                    utils.showNotification(utils.getString(30092))
                     utils.log(utils.getString(30092))
+                    backup_success = False
 
             # reset remote and xbmc vfs
             self.xbmc_vfs.set_root("special://home/")
             self.remote_vfs.set_root(orig_base_path)
 
-            if(utils.getSettingBool("compress_backups")):
+            if(compressing):
                 fileManager = FileManager(self.xbmc_vfs)
 
                 # send the zip file to the real remote vfs
                 zip_name = os.path.join(self.ZIP_TEMP_PATH, self.remote_vfs.root_path[:-1] + ".zip")
                 self.remote_vfs.cleanup()
-                self.xbmc_vfs.rename(os.path.join(self.ZIP_TEMP_PATH, "xbmc_backup_temp.zip"), zip_name)
-                fileManager.addFile(zip_name)
+                renamed = self.xbmc_vfs.rename(
+                    os.path.join(self.ZIP_TEMP_PATH, "xbmc_backup_temp.zip"),
+                    zip_name)
+                if not renamed:
+                    backup_success = False
+                else:
+                    try:
+                        verify_zip_archive(
+                            zip_name, check_cancel=self.progressBar.checkCancel)
+                    except ArchiveValidationError as error:
+                        utils.log('Local ZIP verification failed: %s' % error,
+                                  xbmc.LOGWARNING)
+                        backup_success = False
+                    fileManager.addFile(zip_name)
 
                 # set root to data dir home and reset remote
                 self.xbmc_vfs.set_root(self.ZIP_TEMP_PATH)
                 self.remote_vfs = self.saved_remote_vfs
 
                 # update the amount to transfer
-                self.transferSize = fileManager.fileSize()
-                self.transferLeft = self.transferSize
-                fileCopied = self._copyFiles(fileManager.getFiles(), self.xbmc_vfs, self.remote_vfs)
-
-                if(not fileCopied):
-                    # zip archive copy filed, inform the user
-                    shouldContinue = xbmcgui.Dialog().ok(utils.getString(30089), '%s\n%s' % (utils.getString(30090), utils.getString(30091)))
+                if backup_success:
+                    remote_zip = (self.remote_vfs.root_path +
+                                  os.path.basename(zip_name))
+                    if self.remote_vfs.exists(remote_zip):
+                        utils.log('Backup ZIP already exists: ' + remote_zip,
+                                  xbmc.LOGWARNING)
+                        backup_success = False
+                    else:
+                        remote_artifact = remote_zip
+                        self._active_artifact = remote_zip
+                        self._active_artifact_compressed = True
+                if backup_success:
+                    self.transferSize = fileManager.fileSize()
+                    self.transferLeft = self.transferSize
+                    fileCopied = self._copyFiles(
+                        fileManager.getFiles(), self.xbmc_vfs,
+                        self.remote_vfs)
+                    backup_success = bool(fileCopied)
+                if backup_success:
+                    try:
+                        self._verifyCompressedUpload(zip_name, remote_zip)
+                    except ArchiveValidationError as error:
+                        utils.log('Uploaded ZIP verification failed: %s' % error,
+                                  xbmc.LOGWARNING)
+                        backup_success = False
 
                 # delete the temp zip file
-                self.xbmc_vfs.rmfile(zip_name)
+                if renamed:
+                    self.xbmc_vfs.rmfile(zip_name)
+            elif backup_success:
+                try:
+                    self._verifyFolderBackup(orig_base_path)
+                except ArchiveValidationError as error:
+                    utils.log('Folder backup verification failed: %s' % error,
+                              xbmc.LOGWARNING)
+                    backup_success = False
 
-            # remove old backups
-            self._rotateBackups()
+            backup_success = self._finalizeBackup(
+                backup_success, remote_artifact, compressing)
 
-            # close any files
             self._closeVFS()
+            return backup_success
+
+        return False
 
     def restore(self, progressOverride=False, selectedSets=None):
         shouldContinue = self._setupVFS(self.Restore, progressOverride)
@@ -370,7 +448,7 @@ class XbmcBackup:
                 self.saved_remote_vfs = self.remote_vfs
                 self.remote_vfs = ZipFileSystem(zip_path, "w")
 
-            self.remote_vfs.set_root(self.remote_vfs.root_path + time.strftime("%Y%m%d%H%M") + utils.getSetting('backup_suffix').strip() + "/")
+            self.remote_vfs.set_root(self.remote_vfs.root_path + time.strftime("%Y%m%d%H%M%S") + utils.getSetting('backup_suffix').strip() + "/")
             progressBarTitle = progressBarTitle + utils.getString(30023) + ": " + utils.getString(30016)
         elif(mode == self.Restore and self.restore_point is not None and self.remote_vfs.root_path != ''):
             if(self.restore_point.split('.')[-1] != 'zip'):
@@ -388,18 +466,30 @@ class XbmcBackup:
         # setup the progress bar
         self.progressBar = BackupProgressBar(progressOverride)
         self.progressBar.create(progressBarTitle, utils.getString(30049) + "......")
+        self._vfs_closed = False
 
         # if we made it this far we're good
         return True
 
     def _closeVFS(self):
-        self.xbmc_vfs.cleanup()
-        self.remote_vfs.cleanup()
-        self.progressBar.close()
-
-        # reset the window setting
-        window = xbmcgui.Window(10000)
-        window.setProperty(utils.__addon_id__ + ".running", "")
+        if self._vfs_closed:
+            return
+        self._vfs_closed = True
+        for cleanup in (
+                self.xbmc_vfs.cleanup,
+                self.remote_vfs.cleanup,
+                self.progressBar.close):
+            try:
+                cleanup()
+            except Exception as error:
+                utils.log('Backup cleanup failed: %s' % error,
+                          xbmc.LOGWARNING)
+        try:
+            window = xbmcgui.Window(10000)
+            window.setProperty(utils.__addon_id__ + ".running", "")
+        except Exception as error:
+            utils.log('Unable to clear backup running state: %s' % error,
+                      xbmc.LOGWARNING)
 
     def _copyFiles(self, fileList, source, dest):
         result = True
@@ -412,7 +502,10 @@ class XbmcBackup:
             dest.mkdir(dest.root_path)
 
         for aFile in fileList:
-            if(not self.progressBar.checkCancel()):
+            if(self.progressBar.checkCancel()):
+                result = False
+                break
+            else:
                 if(utils.getSettingBool('verbose_logging')):
                     utils.log('Writing file: ' + aFile['file'])
 
@@ -546,21 +639,62 @@ class XbmcBackup:
                 remove_num = 0
 
                 # update the progress bar if it is available
-                while(remove_num < (len(dirs) - total_backups) and not self.progressBar.checkCancel()):
+                while(remove_num < (len(dirs) - total_backups)):
+                    if self.progressBar.checkCancel():
+                        utils.log('Backup retention cancelled', xbmc.LOGWARNING)
+                        return False
                     self._updateProgress(utils.getString(30054) + " " + dirs[remove_num][1])
                     utils.log("Removing backup " + dirs[remove_num][0])
 
                     if(dirs[remove_num][0].split('.')[-1] == 'zip'):
                         # this is a file, remove it that way
-                        self.remote_vfs.rmfile(self.remote_vfs.clean_path(self.remote_base_path) + dirs[remove_num][0])
+                        removed = self.remote_vfs.rmfile(self.remote_vfs.clean_path(self.remote_base_path) + dirs[remove_num][0])
                     else:
-                        self.remote_vfs.rmdir(self.remote_vfs.clean_path(self.remote_base_path) + dirs[remove_num][0] + "/")
+                        removed = self.remote_vfs.rmdir(self.remote_vfs.clean_path(self.remote_base_path) + dirs[remove_num][0] + "/")
+
+                    if not removed:
+                        utils.log('Backup retention deletion failed',
+                                  xbmc.LOGWARNING)
+                        return False
 
                     remove_num = remove_num + 1
+        return True
+
+    def _finalizeBackup(self, success, artifact_path, compressed):
+        if success:
+            self._active_artifact = None
+            if self._rotateBackups() is False:
+                utils.showNotification(utils.getString(30092))
+                return False
+            return True
+        if artifact_path:
+            if compressed:
+                self.remote_vfs.rmfile(artifact_path)
+            else:
+                self.remote_vfs.rmdir(artifact_path)
+        self._active_artifact = None
+        utils.showNotification(utils.getString(30092))
+        return False
+
+    def _discardActiveArtifact(self):
+        artifact = getattr(self, '_active_artifact', None)
+        if not artifact or getattr(self, 'remote_vfs', None) is None:
+            return
+        try:
+            if getattr(self, '_active_artifact_compressed', False):
+                self.remote_vfs.rmfile(artifact)
+            else:
+                self.remote_vfs.rmdir(artifact)
+        except Exception as error:
+            utils.log('Unable to remove failed backup artifact: %s' % error,
+                      xbmc.LOGWARNING)
+        finally:
+            self._active_artifact = None
 
     def _createValidationFile(self, dirList):
         gui_settings = GuiSettingsManager()
-        valInfo = build_manifest(dirList, self._hashFile, {
+        valInfo = build_manifest(dirList, lambda path: self._hashFile(
+            path, cancellable=True), {
             "name": "Backup Pro Manifest",
             "created_utc": datetime.utcnow().strftime('%Y-%m-%dT%H:%M:%SZ'),
             "addon_id": utils.__addon_id__,
@@ -571,7 +705,8 @@ class XbmcBackup:
             "system_settings": gui_settings.backup(),
             "addons": gui_settings.list_addons(),
             "plan": self.backup_plan,
-        })
+        }, check_cancel=self.progressBar.checkCancel)
+        self.backup_manifest = valInfo
 
         local_manifest = xbmcvfs.translatePath(
             utils.data_dir() + MANIFEST_NAME)
@@ -579,6 +714,7 @@ class XbmcBackup:
         vFile.write(json.dumps(valInfo, sort_keys=True, separators=(',', ':')))
         vFile.write("")
         vFile.close()
+        self.backup_manifest_file_hash = self._hashFile(local_manifest)
 
         success = self._copyFile(
             self.xbmc_vfs, self.remote_vfs, local_manifest,
@@ -636,9 +772,66 @@ class XbmcBackup:
 
         return result
 
-    def _hashFile(self, path):
+    def _hashFile(self, path, cancellable=False):
         with xbmcvfs.File(xbmcvfs.translatePath(path), 'r') as source:
-            return sha256_reader(source.readBytes)
+            check_cancel = (self.progressBar.checkCancel
+                            if cancellable else None)
+            return sha256_reader(
+                source.readBytes, check_cancel=check_cancel)
+
+    def _hashVfsFile(self, vfs, path, cancellable=False):
+        if not isinstance(vfs, DropboxFileSystem):
+            return self._hashFile(path, cancellable=cancellable)
+
+        local_copy = xbmcvfs.translatePath(
+            utils.data_dir() + 'backup-pro.readback.tmp')
+        try:
+            if xbmcvfs.exists(local_copy):
+                xbmcvfs.delete(local_copy)
+            if not vfs.get_file(path, local_copy):
+                raise IOError('remote download failed')
+            if cancellable and self.progressBar.checkCancel():
+                raise ArchiveValidationError('remote read-back cancelled')
+            return self._hashFile(local_copy, cancellable=cancellable)
+        finally:
+            if xbmcvfs.exists(local_copy):
+                xbmcvfs.delete(local_copy)
+
+    def _verifyFolderBackup(self, backup_root):
+        root = self.remote_vfs.clean_path(backup_root)
+        try:
+            manifest_hash = self._hashVfsFile(
+                self.remote_vfs, root + MANIFEST_NAME, cancellable=True)
+        except Exception as error:
+            raise ArchiveValidationError(
+                'unable to read back published manifest: %s' % error)
+        if manifest_hash != self.backup_manifest_file_hash:
+            raise ArchiveValidationError(
+                'published manifest does not match the local manifest')
+        result = verify_manifest_files(
+            self.backup_manifest,
+            lambda relative: self._hashVfsFile(
+                self.remote_vfs, root + relative, cancellable=True),
+            check_cancel=self.progressBar.checkCancel)
+        utils.log('Verified folder backup: %s' % json.dumps(result))
+        return result
+
+    def _verifyCompressedUpload(self, local_zip, remote_zip):
+        try:
+            local_checksum, local_size = self._hashFile(
+                local_zip, cancellable=True)
+            remote_checksum, remote_size = self._hashVfsFile(
+                self.remote_vfs, remote_zip, cancellable=True)
+        except Exception as error:
+            raise ArchiveValidationError(
+                'unable to read back uploaded ZIP: %s' % error)
+        if remote_size != local_size or remote_checksum != local_checksum:
+            raise ArchiveValidationError(
+                'uploaded ZIP does not match the verified local archive')
+        result = {'file_count': 1, 'total_bytes': local_size}
+        utils.log('Verified compressed backup upload: %s' % json.dumps(
+            result, sort_keys=True))
+        return result
 
     def _createResumeBackupFile(self):
         with xbmcvfs.File(xbmcvfs.translatePath(utils.data_dir() + "resume.txt"), 'w') as f:
